+++ examples/daily_task.rs
@@ -15,5 +15,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .start_boundary(&Local::now().to_rfc3339())?
         .build()?
         .register("TaskName", TaskCreationFlags::CreateOrUpdate as i32)?;
+
     Ok(())
 }