+++ src/schedule.rs
@@ -1,13 +1,15 @@
 use windows::core::BSTR;
+use windows::Win32::System::Com::VARIANT;
 use windows::Win32::System::TaskScheduler::{
     IActionCollection, IRegistrationInfo, ITaskDefinition, ITaskFolder, ITaskService,
     ITaskSettings, ITrigger, ITriggerCollection, TASK_LOGON_INTERACTIVE_TOKEN,
 };
 
 #[derive(Debug, PartialEq)]
-/// A schedule is created by a [schedule builder](crate::schedule_builder). Once created, the 
+/// A schedule is created by a [schedule builder](crate::schedule_builder). Once created, the
 /// Schedule can be registered with the Windows Task Scheduler.
 pub struct Schedule {
+    pub(crate) task_folder: ITaskFolder,
     pub(crate) actions: IActionCollection,
     pub(crate) force_start_boundary: bool,
     pub(crate) registration_info: IRegistrationInfo,
@@ -23,16 +25,15 @@ impl Schedule {
     /// Registers the schedule. Flags can be set by using the [TaskCreationFlags](crate::enums::TaskCreationFlags) enum.
     pub fn register(self, task_name: &str, flags: i32) -> Result<(), Box<dyn std::error::Error>> {
         unsafe {
-            let folder: ITaskFolder = self.task_service.GetFolder(&"\\".into())?;
-            folder.RegisterTaskDefinition(
+            self.task_folder.RegisterTaskDefinition(
                 &BSTR::from(task_name),
                 &self.task_definition,
                 flags,
                 // TODO allow user to specify creds
-                None,
-                None,
+                VARIANT::default(),
+                VARIANT::default(),
                 TASK_LOGON_INTERACTIVE_TOKEN,
-                None,
+                VARIANT::default(),
             )?;
         }